requirements.txt
gunicorn==20.1.0
requests==2.28.1
psycopg2>=2.8
Django==4.0.6
djangorestframework==3.13.1
django-filter==22.1
Markdown==3.3.7
django-environ==0.9.0
django-extensions==3.2.0
django-heroku==0.3.1
